+++ packages/nimble-components/src/label-provider/base/styles.ts
@@ -1,0 +1,6 @@
+import { css } from '@microsoft/fast-element';
+import { display } from '../../utilities/style/display';
+
+export const styles = css`
+    ${display('none')}
+`;