+++ packages/nimble-components/src/select/styles.ts
@@ -6,9 +6,11 @@ import {
     borderRgbPartialColor,
     borderWidth,
     controlHeight,
+    controlSlimHeight,
     fillHoverColor,
     fillHoverSelectedColor,
     fillSelectedColor,
+    iconColor,
     mediumPadding,
     placeholderFontColor,
     smallPadding
@@ -37,8 +39,14 @@ export const styles = css`
         order: 1;
     }
 
-    [part='indicator'] {
+    .clear-button {
         order: 3;
+        height: ${controlSlimHeight};
+        margin-left: ${smallPadding};
+    }
+
+    [part='indicator'] {
+        order: 4;
     }
 
     .error-icon {
@@ -95,6 +103,7 @@ export const styles = css`
 
     .filter-icon {
         padding-left: ${smallPadding};
+        ${iconColor.cssCustomProperty}: ${placeholderFontColor};
     }
 
     .filter-input {